imrove documentation
