+++ src/utils/CMakeLists.txt
@@ -5,4 +5,6 @@ target_sources(idefix
   PUBLIC ${CMAKE_CURRENT_LIST_DIR}/dumpImage.cpp
   PUBLIC ${CMAKE_CURRENT_LIST_DIR}/dumpImage.hpp
   PUBLIC ${CMAKE_CURRENT_LIST_DIR}/lookupTable.hpp
+  PUBLIC ${CMAKE_CURRENT_LIST_DIR}/column.cpp
+  PUBLIC ${CMAKE_CURRENT_LIST_DIR}/column.hpp
   )
